Use "python-ldap" < 3.0.0b1.
"python-ldap" 3.0.0b1 released on 2017-12-04 changed the way it handles text and bytes (see https://python-ldap.readthedocs.io/en/latest/bytes_mode.html#what-s-text-and-what-s-bytes) which caused some tests to fail.

As a workaround, we need to use an older release of "python-ldap" until we can thoroughly investigate the implications of the changes in "python-ldap".